refactor(coderd/x/chatd): own provider option construction in one function
Provider option conversion and reasoning effort injection both create OpenAI
option structs, so each call site had to pair them correctly and pick the same
transport. ProviderOptionsForCall now owns both steps, and the two helpers are
unexported.
