+++ coderd/x/chatd/chatprovider/chatprovider_test.go
@@ -387,14 +387,16 @@ func TestAnthropicThinkingDisplayFromChat(t *testing.T) {
 	}
 }
 
-func TestProviderOptionsFromChatModelConfig_AnthropicThinkingDisplay(t *testing.T) {
+func TestProviderOptionsForCall_AnthropicThinkingDisplay(t *testing.T) {
 	t.Parallel()
 
-	providerOptions := chatprovider.ProviderOptionsFromChatModelConfig(chatprovider.Model{}, &codersdk.ChatModelProviderOptions{
-		Anthropic: &codersdk.ChatModelAnthropicProviderOptions{
-			ThinkingDisplay: ptr.Ref(" SUMMARIZED "),
+	providerOptions := chatprovider.ProviderOptionsForCall(chatprovider.Model{}, codersdk.ChatModelCallConfig{
+		ProviderOptions: &codersdk.ChatModelProviderOptions{
+			Anthropic: &codersdk.ChatModelAnthropicProviderOptions{
+				ThinkingDisplay: ptr.Ref(" SUMMARIZED "),
+			},
 		},
-	})
+	}, nil)
 
 	require.NotNil(t, providerOptions)
 	anthropicOptions, ok := providerOptions[fantasyanthropic.Name].(*fantasyanthropic.ProviderOptions)
