+++ coderd/x/chatd/generation_preparer.go
@@ -542,19 +542,7 @@ func (server *Server) prepareGeneration(
 	if chat.LastReasoningEffort.Valid {
 		requestedEffort = new(string(chat.LastReasoningEffort.ChatReasoningEffort))
 	}
-	reasoningEffort := chatprovider.ResolveReasoningEffort(
-		requestedEffort,
-		callConfig.ReasoningEffort,
-	)
-	providerOptions := chatprovider.ProviderOptionsFromChatModelConfig(
-		model,
-		callConfig.ProviderOptions,
-	)
-	providerOptions = chatprovider.ApplyReasoningEffort(
-		model,
-		providerOptions,
-		reasoningEffort,
-	)
+	providerOptions := chatprovider.ProviderOptionsForCall(model, callConfig, requestedEffort)
 
 	activeToolNames := activeToolNamesForTurn(tools, currentPlanMode, chat.ParentChatID, approvedPlanMCPConfigIDs)
 	if isExploreSubagent {
